fix: move db to adapters package
