+++ internal/finance/adapters/db/offce_repo_test.go
@@ -8,8 +8,8 @@ import (
 
 	"sumni-finance-backend/internal/common"
 	"sumni-finance-backend/internal/common/testutils"
+	"sumni-finance-backend/internal/finance/adapters/db"
 	"sumni-finance-backend/internal/finance/app/models"
-	"sumni-finance-backend/internal/finance/db"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/require"
