+++ internal/finance/adapters/db/queries/fund_provider.sql
