+++ internal/finance/adapters/db/queries/wallet.sql
