+++ internal/finance/adapters/db/sqlc.yaml
